pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "test task for absolut"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "django-filter>=24.0",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "python-magic>=0.4.27",
    "django-cors-headers>=4.7.0",
]
